refactor: isolate legacy /api/counters endpoints with deprecation logging.
